files structure refactoring
