+++ lib/ui/screen/add_sight_screen/widget/category_selector.dart
