+++ lib/ui/widgets/sight_card.dart
